feat: enhance type annotations and improve code clarity across multiple files

+++ src/Traits/WithFileSize.php
@@ -25,7 +25,7 @@ trait WithFileSize
         $units = ['B', 'KB', 'MB', 'GB', 'TB'];
 
         $bytes = max($bytes, 0);
-        $pow = floor(($bytes ? log($bytes) : 0) / log(1024));
+        $pow = (int) floor(($bytes ? log($bytes) : 0) / log(1024));
         $pow = min($pow, count($units) - 1);
 
         $bytes /= pow(1024, $pow);
